FIX: refactor comments/profile folders
