+++ src/Components/Profile/index.tsx
@@ -4,7 +4,7 @@ import { User } from '../../types/typeSection';
 const ProfileSection: React.FC<{ user: User }> = ({ user }) => {
   return (
     <div className='w-full mb-3'>
-      <section className='w-80 h-96 mx-auto bg-[#20354b] rounded-2xl px-8 py-6 shadow-lg'>
+      <section className='w-72 sm:w-80 h-96 mx-auto bg-[#20354b] rounded-2xl px-8 py-6 shadow-lg'>
         <div className='mt-6 w-fit mx-auto'>
           <img src={user?.picture || ''} alt='' referrerPolicy='no-referrer' />
         </div>
